Initial commit of the MICROS project.
The project contains files to build a micro OS that shows a message on the screen. The project is manipulated from the makefile in the directory. Use "make build" to create the output files or simple "make run" to start qemu and run the OS in i386 emulation mode. If you are starting qemu from the command line the keys to exit the program is ESC-2 and when 'q' at the qemu prompt.

The built kernel can be run from Windows but must be built with custom tooling from a Linux machine. The tools include a custom version of GCC and binutils used to cross compile the binaries. The tools will be uploaded in the repo soon but for now a branch will be created with the build kernel and binaries.