Build of the binaries of the kernel.
Those have been built from a Linux machine that contains my tools that are custom made for this kernel.

See instruction at http://wiki.osdev.org/GCC_Cross-Compiler to see how to build the tools.

To run the kernel in qemu from Linux run 'run.sh' and from windows the script is 'runWin.cmd' provided that qemu is installed and that the batch script can find the executable at the same place as on my computer.