Initial customization of the project
Organization of the makefile into categories. Added a debug target to start the program rather than calling the run.sh script. Removed the run script for Windows because its useless to start the emulator on Windows.

Added a gdb init script so that when entering 'gdb' simply, will enter and break at the kernel_main function.

Replaced the Hello World with a color test pattern.

Since development has started I added a LICENSE file to enable the MIT license copyright.