Logging and basic Memory Management
 - Added methods to send strings to the kernel COM logging port.

- Added "very basic" memory management. The memory manager currently
starts at the 45MB mark and allocates from there. No free() for now.
But hey, it works.

 - Added an -O0 switch to the GCC since -Og did not seem to remove all
optimizations. The compiler was optimizing some code and values were
being optmized out when stepping through the debugger.

 - Added a basic implementation of the serial ReadByte method but
currently untested and will block as long as the bytesToRead mark is
not reached.

 - Started implementing a function to get the string representation of
a number. This will be used in the printf function implementation.