Keyboard driver basic implementation.
 - Created a common.h header file that will contain utilities most code
will use. For now it only contains an empty Debugger function that is a
hook for gdb to break into. Using this, we can place Debugger() calls
in the code and gdb will break on those, this is easier that constantly
editing .gdbinit to add functions to the break list. It has the
disadvantage to break "during" the function rather that acting like a
break point so we always need to do a "next step" before returning to
the calling scope.

 - The keyboard driver is very basic for now. The driver keeps a list
of all keys and the state pressed or not. Each entry in the key_states
array is the state of each key. The value of the keycode is the index
into the key_states list. If the key is pressed the value will be one
and when the key gets released the driver sets the field to 0.

 - The keyboard_state structure has been reworked and contains a
pointer to the keyStates array to manually check if a key is pressed.
The method IsKeyDown does the same thing.