Memory Manager second implementation.
 - The new memory manager has been upgraded to support freeing memory.
The previous one just incremented the next allocation mark and returned
the current value as a pointer. This had the limitation of not being
able to free memory.

 - The second version is a "pooled" memory manager. The MM maintains 3
pools (the 4th pool is a secret pool). The small pool is reserved for
allocations smaller than 256 bytes and contains a fixed amount of 10k
elements. The second pool is the page pool, contains 610 elements of
4kb size or less. The third pool is for allocations of 32kb and can
contain 160 elements.

 - The first and second pool take up 2.5 MB and the third pool takes
5MB.

 - The MM scans the list for a free allocation unit and marks it as
unavailable before returning a pointer to the memory location pointed
by the unit. There is no implementation right now to deal with bad
cases like full memory or a pool being all taken.

 - This version is only an incremental improvement to allow memory
frees. It is far from being fully workable.

 - The 4th pool is the 'crazy pool'. It doesn't exist. Allocations
bigger than 32kb can go fuck themselves. For now.