Kernel milliseconds timing upgrade
 - Changed the timer divider to allow a tick every ms or so. Not
precise for now due to ticks counting mechanism that does not increment
exactly at each 1000 counts.

 - The screen shows an incrementing timer of milliseconds to show the
code working.

 - This code is doing a lot of number printing on the screen, which
calls sprintf a lot and it needs to allocate a bunch of memory in order
to convert the int to strings. This now works due to the new memory
manager that can deallocate memory.