Basic interrupt error handling
 - Added an error.h file with a variable named panic that can be
switched to TRUE so the computer will stop cycles and exit the loop.

 - Currently, when doing a divide by 0, an interrupt will be launched.
I catch the interrupt and increment EIP by 1 so the instruction right
next of the bad division will execute. Not a good implementation
because I need to GTFO immediately into a kernel panic mode.

 - This implementation fixes the problem of an "unhandled" interrupt.
Because errors will push the address of the faulting instruction into
the stack. When jumping out of the interrupt, EIP will get popped back
the address of the faulting instruction and it will be executed again.
Generating another interrupt. The way to fix this is to point EIP to
another value, in this case the next instruction. In the future the
kernel will jump to kernel_panic method. Otherwise the kernel will
never return and will flood the log with error messages.

 - Changed a compile option from -g to -g3 to increase debug info
generated. g3 will keep the #define variables instead of converting them
directly to lvalues. This allows to get the #define value using "print
large_pool_size" to get 160.