Error handling initial implementation
 - Error file contains procecures for handling CPU faults as they are
raised as interrupts during the execution. idt.c will check the
interrupt code and call the appropriate function.

 - Currently handing Division by 0 and General Protection faults by
calling error() which jumps out of the kernel execution loop.

 - boot.s renamed to boot.asm since I was refactoring it for this
commit.

 - boot.asm's labels were renamed without leading periods to make them
global (rather than "local" labels relative to the _start function.
This is done to make anyone able to jump to the error: label in order
to hang the CPU directly.