Finalized the commandline parser.
The commandline parser takes a command-line string and returns a list
of arguments. Two types of arguments can be returned. Simple arguments
are bunch of letters between separating spaces and switch arguments are
a two-part name-value pair. The name part of a switch argument must
contain a leading '-' and then the name of the argument followed by a
space. After the space the value of the argument is spelled out.
Arguments may not contain spaces in their values (simple arguments are
basically switch arguments without names) because spaces separate the
arguments.

The parser is a state machine going through each character in turn and
building an array of strings of each arguments.