GDB debugging macro creation : jumpd.
With the GDB jumpd function, we can jump to any line in the current
code file and break at that position. This can help to restart a
function execution without having to recompile and restart the GDB
session. It does NOT 'rewind' the execution though so exercise caution
when re-executing code that increments globals.