Common and error headers rework.
Common.h is now an unshielded header used to include the most common
used headers.

Added a new header types.h to put the BOOL types and other new types.

Added a PanicQuit function used to show a message on the framebuffer
and hang the processor. It is used in the ASSERT macro if the assert
check fails.

The Debugger function has been moved to the error.h because I need
another place for it.