Rework of the copy functions.
The old copy function was crap. The size of the pointer was always 1
so it was only copying 1 byte. 2 new functions replace the old one.

The c standard memcpy function takes a count of bytes to copy. And the
replaced kmKernelCopy now needs to look in the allocation pools to
check the size of the pointed structure.

If we try to kmKernelCopy with pointers that are not from the stack it
won't work because the data is not allocated somewhere where we can
verify its size. Still working on it.

Now the pools track the data size that we were passing anyway but I
forgot to assign it to the alloc struct.