Function to enumerate all PCI devices.
Added a pci_device.valid_device member to indicate if the device is not
present on the computer. This is done so we can enumerate all possible
values and then check if it's valid to add it in a list. This is
because the only way to check for the presense of a PCI device is to
query it and we don't want to ASSERT if we're wrong.

This version stores the devices in a vector_create_size(15) because
vector resize does not currently work. My testbed has 12 devices so it
never gets resized.