New filesystem files and ATA upgrade.
Starting the implementation of the filesystem module.

Added functions in the ATA driver to get a status field. This field is
more relevant to outside modules, to check if the driver is good to go
or not.

Replaced the sectors variable types to 64 bits to support a future
upgrade to LBA 48 addressing.