ATA driver bugfixing. R/W IO and refactor.
Fixed the Read and Write methods in the case where more than 256 bytes
were written. The functions were leaving the DRQ wait loop after 255
bytes written because the disk goes busy for a bit in-between each
sector write. Added a busy wait so the driver keeps looping. I am NOT
checking for DATA_REQUEST in the loop because during the last iteration
the DATA REQUEST will be false and will get stuck in the loop forever.
Using the 'for' IO method means that there will always be a specific
number of writes and checking DRQ is pointless.

Added a write_test_sectors method so write some data to the disk to be
read back to test if the IO mechanisms are working. Need to work on
reading back the memory and confirming if it is indeed working well.

Added stuck IO checking using a flag that is defined if the _DEBUG
profile is being run. Build profiles are enabled in the new micros.h
header that each other files should always include. This will contain
kernel global definitions.