Fix for the wrong read/write.
Problem was that the wrong amount of bits were selected, the LBA bits
must be sent in 8 bits quantities except in the first register where it
is only the high 4 bits. The code was only selecting 4 bit quantities
so some number of sectors to write that didn't use the top 4 bits of
their words.