Filesystem read and write implementation.
Correctly determining the sector to write to from a disk address. The
function uses a ceil and divide to find the offset within a sector
where the IO operation starts.

Finished the implementation of the read and write data functions. They
use too many mem copy operations so it will be slow and inefficient for
now. It's using PIO mode anyway so it was always going to be
inefficient.