ext4 refactor over to EZFS.
New filesystem is simple and will serve as the first working filesystem
in MICROS. The filesystem will keep files in one contiguous block and
will have to relocate if the file grows.

Initial implementation is only a few of the formatting functions but I
expanded on the required functionality from the filesystem layer.