ATA write_data fix.
Fixed the implementation of write_data that was not reading the
existing data from the disk in order to write the old data + new data
back. Also fixed the fact that the function was not aligning the buffer
to the correct sector offset.

Changed the signature of ezfs_read_file function to return the buffer
as an out parameter.