Added a sortable file allocation table.
This table is a pointer to pointers of file_allocations. This means
that I can sort the upper pointers based on disk address and use that
list instead when looking between files to find some free space. This
is because, to look between two files, I need to compare the current
iterated file to the [current +1] file and compare the size diff
between them. For that to happen then need to be sorted by size.

Still working out the quirks of the new system. Need to make a test
bench for deleting files in order to make data holes between files to
test the sorting function correctly.