Error in handling kernel module addresses.
When determining the addresses for the kernel modules, the error was to
use the reference of the kernel_info struct. That reference was coming
from a statically assigned variable and so was overflowing on the
memory manager pool's pointers.

Also, temporarily using the memory manager to allocate the kernel
structs because even with the above the the memory was still
overflowing.