Keyboard driver upgrade
Added support for extended scan codes. This will get the arrow keys when they are pressed. They are currently outputting weird other keys but work in progress.

Fixed a bug in ksh where the first line would get assigned the address of the "\0" string.

Added a write outputting command. Still work in progress.