Fixed EZFS functions.
Fixed the ezfs_find_free_space by returning the start of the data area rather than the data address of an unexistant file.

Added a few missing casts for the mallocs.

When reading a file, checking if found is NULL, rather than false.

Checking if a file was allocated before returning it by name.