Wrapping up first phase of heap MM
Switching to load the memory manager heap structure at 6MB rather than 5 because it was overwriting the IDT and GDT tables. This is a temporary solution since we will find a new home for those two tables.

Added parentheses to the MM defines so that their conditions do not affect operator precedence outside the define.

Added handling of more cases of malloc.

Commented out much of the old setup calls in kernel to test the changes without clashes.