Fixed buffer overflow in the ATA Driver.
Wrrong number of bytes was requested, the number requested was okay for a single sector but needs to be multiplied by the number of sector that we need to read.

Also added a missing free() in the test_sector function, not really a problem since this function is not used but still, proper now.