Memory allocation debug feature
Added a debug feature called Alloc Canaries. These are special bytes appended to the end of a malloc'ed range. These bytes are checked when the alloc is freed or by demand, if these bytes have changed it means that a buffer overflow happened and overwrote those bytes somehow.

It doesn't detect if the buffer was changed much further than the canary bytes though. Also does not stop callers from just reading past the buffer.

This feature is enabled by the define MM_ENABLE_HEAP_ALLOC_CANARY since this debug feature adds bytes to each allocation and is more expensive to malloc/free.