Page Frame management structure and functions.
Page Frames are the 4KB segments in physical RAM. Those sections are
mapped to by the page allocator to a virtual page. The structure
page_frame_map occupies 1 MB (256 pages) and each byte of that
structure maps to a page frame. Each element of the structure stores
flags so we can keep track of what area of memory is used.

The page_frame_map should always be mapped in memory.