Added stack allocated sting functions for PA.
The PageAllocator needs to output to the log file and/or to the screen,
the string formatting functions used are using malloc to get memory
from the heap but at the time of the PA init the kernel heap is not
created yet. The new functions use stack allocated memory with some
starting sizes to fill the allocations needed.