Reworked the page finding algorithm + docs.
Added a Page Allocator module structure, this structure is used to keep
some info. This info so far contains minimum and maximum page number
where allocations will be pulled from. So far the structure is
allocated in the data segment of the kernel and I made a pointer to it.
Sometimes soon the struct will be allocated in the kernel bloc.

Also fixed the way to find free pages, directories are allocated if
need be (correctly this time).