Fixed memory manager alloc pointers.
Fixing the error of taking the address of a pointer when doing a malloc. Also fixed the flags assignment that was forgotten.

Also now, when doing a MEM_CHECKED allocation the size is precalculated to take into account the canary bytes so the correct size is used to calculate available space. Before this the size was +5 bytes but the space that was found in the heap wasn't taking into account the bigger size.