Manual page allocation for heap manager.
Turned the load on demand for pages and having the memory manager
allocate all its pages in advance.

Had to remove some includes in the paging module, this is because
paging.h is including common.h and it contains the memory manager
header, that in turn creates a circular reference back to paging.h.
Paging.h now only includes the headers it needs.