Enabled the IRQ bytes transfer to new stack.
Missed a +4 increment in the commented code. Now the EIP, CS and EFLAGS
are transfered to new stack. IRQ switches to new stack at the last
minute and lets the IRET pops those values back.

The only things left is to save the interrupted task and replace the
stack correctly.