Setting the CPU to execute task1 instead of sleep.
Added a +16 increment to the ESP we save to the task structure to
offset the 3 values pushed by the interrupt. Might find a cleaner way
than just bumping it 16 bytes but this fix avoids the task stack from
underflowing the stack.

Setting GDB to show the ASM and code panels by default to help
debugging task switching.