Task switch corrections
When activating a thread, if EIP is 0 (a way to check if it never
executed any code yet) we copy the entry addr into EIP. Also need to
set it to 0 when creating a thread instead of a test value.

Incrementing the current index when we get the next thread in line.

When activating a thread, read the correct EIP value instead of the
entryAddr value.