Task Switch algorithm change.
This introduces a new way to get a task's registers and state. The new
way is calling ks_save_fix_registers (new name at some point) as SOON
as possible during the ks_suspend call, dumps the registers and walk
the stack back up 2 frames (ks_save_fix_registers -> ks_suspend ->
[target calling site]) and gets the EIP and ESP register values from
the top (end) of that stack frame.

This will enable the task to be restored to the next instruction right
after ks_suspend, avoiding the clunky second-pass of the ks_get_current
function. To do this, calling ks_save_fix_registers must be the first
thing we do during ks_suspend to avoid having a register clobbered.
Currently, only ESP, EBP and EIP are doing to be changed and we can
restore those with the stack info. Currently using a hardcoded byte
increase for ESP instead of dynamically walking the frames back up.
Also not done with all the register copying and returning a struct with
the new function.

Added an example of the assembly required from the godbolt explorer
(upload to https://godbolt.org/) to see output assembly.