Working task suspension mechanism !
Finally got it working.

This implementation is using a 2-stage ks_suspend implementation. The
first stage saves the registers on the stack, this part can be done in
assembly without too much trouble. The second stage can be coded in c
and reads the registers, save them and does the job of scheduling the
next thread. Some documentation is included in for the functions.

This method was used to avoid creating too complex code in assembly and
to make sure that we can restore tasks reliably and without restoring a
task in the middle of the register saving as before. This method wakes
up the task at the next instruction after the call to ks_suspend.