Updated the timer code for the scheduler.
Created a define for the timer to fire every microsecond. The division
does not equal exactly 1us but gets rounded. We avoid doing too much
work in an interrupt going off every microsecond so we update the tick
variable until we get 1000 and increment the mscounter.

Added a function to get the interval of the timer, this is the interval
at wich the timer emits an event and not every time the PIT fires.