Old preempt mechanism migration to the scheduler.
Deleted much of the old task management code in kernel.c. They have
been moved to task.h.

Assembly of the timer stub in charge of checking for task preempt has
the function references changed to the new functions. The new functions
pretty much work the same.

The preempt mechanism is disabled by having ks_should_preempt_current
return false always. Will test the code for logical flaws after.

The ks_get_next_thread was changed to return the index of the next task
by parameter. This is due to a flaw in the work that the function was
doing. The functions should only GET the next thread object and not set
it as current in the scheduler structure. The caller should do that
work but it also needs to set a variable indicating the index of the
new task so we return it also.