Rework of the module system.
Added a memory_zones.h file that contains the addresses of the zones of
memory that are reserved for a specific purpose. Will work on a
graphical document ref soon.

Added a 1MB zone dedicateds to containing the kernel module structure.
Starting at 1MB for now and placed it at the 9MB mark.

Identity mapping the first 16MB's to allow the zones to work. All the
zones currently live in identity mapped memory. Might change at some
point.

Changes in the page allocator is kinda wonky, reserving 16MB from start
and added 2 more page directories to handle the 2x 4MB additional
ranges to cover the first 16MB. Didn't test a lot, maybe those
directories will get overwritten by someone. Also not sure if the pages
1027 and 1028 are valid to contain a directory.

Fixed a bug in the calculation of free space in the kernel modules zone.