Started to deal with task priorities.
Tasks can have priorities and obviously, tasks with higher priority
will get more CPU time. Currently this type of time distribution is not
implemented but there is a special priority T_PCRITICAL that is meant
for time sensitive tasks.

Critical tasks will get woken from SLEEPING state at the exact time
that they need. Normal threads are only guaranteed to be woken AT or
AFTER their requested time. The way it works is each timer IRQ tick we
check the list of critical tasks to see if anyone is sleeping and is
due now if yes, we preempt the current thread even if his timeout is
not reached yet.

Tasks can be promoted to T_PCRITICAL with the function
ks_criticalize_task, if will ensure that it is hooked in the right
places.

The critical task list is created in order to speed up the process that
takes place each timer IRQ, only the critical list will be scanned and
not the normal tasks. Only if the current thread timeout is reached
that we will check the other normal tasks.

Critical threads get preempted like all other threads. For a thread to
avoid being preempted, it needs to change state to T_UNINTERRUPTIBLE.
Critical threads are meant more for precise sleep wakeup. Good for
device communication and polling.