Documentation and README update.
Updated the TODO priorities, crossed off a few points, added a few more.

Updated the README with some guidance to download, install and run the
project. Still missing a better format for presenting the info but will
do for now. Updated the feature set and added a category for
contributing.