Multiboot info printing.
Printing the multiboot info struct, added a parameter to the memory
manager module to accept the multiboot struct.

Added a new function to print the data, just the addresses. Not
currently dereferencing the data to print the inner structs.

Shortcutting the heap creation so far, only testing code for the moment.