Finalization of the sprintf usage.
Replaced all calls to the old sprintf_1d over to sprintf.

Deleted the "stacksafe" functions because all the functions called by
sprintf are now using the memory from the stack.

Changed the signature of several string functions to allow the passing
of an external char buffer instead of allocating memory from heap.

Added a vnsprintf function to accept a va_list, this allows other
variadic functions to pass variadic arguments, for example the
kWriteLog.

Upgraded the kWritelog to have only one version that accepts a variable
number of arguments.