Better heap management.
The function mm_create_heap now creates a heap structure and is able to
correctly find room for a heap in the memory.

Added more "malloc" methods named vmalloc and vmfree. These methods
target the heap of the current process (currently using task instead,
will fix) and the functions kmalloc are using the kernel heap
automatically.