Work on PCI scanning in advance of the SATA dev.
Added a new device in QEMU for a IDE disk connected to an AHCI bus. It
seems that SATA drives may be able to be used using the existing ATA
driver but I don't know. The new device is there to be recognized by
the PCI scanning routine.

PCI scanning was upgraded to report the names of the Class and Subclass
using a map, still some quirks due to some numbers being
non-sequential. I added some exceptions to deal with this and mark them
as generic or undefined.