Disk Manager initial work.
The disk manager will serve as the abstract system to handle different
kinds of disks in a smooth way. Currently the ATA driver can only
select one drive, throw a kernel panic on the slightest error, has bad
issues with polling and does not track any state information.

I started the work on making the ATA driver sturdier by returning
integer error codes when a problem arises. Introduced a timeout when
spinning to avoid infinite IO STUCK WAITING FOR STATUS messages.

Fixed the ata_identify_device struct that was incorrectly named in a
few places and turned on tight packing to avoid the struct going to a
size 516 bytes, still not sure why the compiler would pack my 512 size
into a 516.

Removed the test SATA disk as it was making my IDE disk tank.