Paging function wrapping
Seems that the compiler started to complain about disablePaging being
declared in another file. Since the enablePaging disablePaging are
implemented in asm and only included in the paging module I'll make
official wrappers in the Page Allocator module so they can be resolved
safely. Also to be able to do checking and stuff.