Struct packing !!!!
Number of hours spent debugging an issue caused by struct packing = 5.

Old total was 3 but I wasn't keeping count. 2 new hours this morning.

Issue was caused by the PCI structure this time, when passing the
pci_controlset to the ahci scan function, the scan function was getting
garbage PCI devices because of an offset of a couple of bytes in the
structure.

I'm sure there is some danger with tight struct packing but for now this
fixes the problem.