More AHCI implementation.
Added memory setup for the AHCI structs. We need to point the commands
to a memory area that we provide. Currently setupping on the 10MB mark
but will be dynamically allocated eventually.

Added some more structure packing, I'm not sure if it fixes the problem
or just moves the problem around.

Some testing calls in kernel.c

Added more macros to manipulate flags.