Found why no keyboard events were raised.
When creating a new thread, the flags register receives the value 0
initially. Which means interrupts are not raised. So the first process
we switch to will get exclusive control with no interrupts.

To fix the issue for now, I added an enable_interrupts at the entry
points of the threads I am using. The flags register gets saved
correctly and will switch tasks.