Added method to check if redraw is needed.
The problem was that the main screen would not check if any of its watched ports changed to trigger a redraw, only the data from the host struct was compared for a redraw. Moved all the checks to a method and we call that every frame. Much cleaner.

Only did the host check so far.