Added new screen to the AHCI tester.
The Main CAP screen shows the CAP byte of the HBA. Added a shortcuts list to the main screen.

Added a new string format to show a binary result based on the argument. It's probably not a standard argument but I'll leave it there for now because it was easy to implement.