Added Docker buildscripts
These build scripts are used to invoke a docker container to make the build of the OS. It can even be used to run the OS and observe the output from the terminal, it also allows input resulting in a fully testable environment without having to install all the OS build dependencies.

Also added PowerShell scripts to use on a Windows machine. The PS scripts use “cd..” To jump back 1 directory but I don’t know if this works on Powershell to set the CWD of the scripts.

The shell scripts work fine so, to hell with Powershell for now.